+++ hyperspy/_components/polynomial.py
@@ -52,9 +52,6 @@ class Polynomial(Expression):
     """
 
     def __init__(self, order=2, module="numexpr", **kwargs):
-        # Not to break scripts once we remove the legacy Polynomial
-        if "legacy" in kwargs:
-            del kwargs["legacy"]
         if order == 0:
             raise ValueError("Polynomial of order 0 is not supported.")
         coeff_list = ['{}'.format(o).zfill(len(list(str(order)))) for o in
@@ -64,7 +61,6 @@ class Polynomial(Expression):
         name = "{} order Polynomial".format(ordinal(order))
         super().__init__(expression=expr, name=name, module=module,
              autodoc=False, **kwargs)
-        self._id_name = "eab91275-88db-4855-917a-cdcbe7209592"
 
     def get_polynomial_order(self):
         return len(self.parameters) - 1
@@ -145,14 +141,12 @@ def convert_to_polynomial(poly_dict):
     """
     Convert the dictionary from the old to the new polynomial definition
     """
-    _logger.info("Converting the polynomial to the new definition")
-    poly_order = poly_dict['order']
+    _logger.info("Converting the polynomial to the new definition.")
     coeff_list = ['{}'.format(o).zfill(len(list(str(poly_dict['order']))))
                   for o in range(poly_dict['order'], -1, -1)]
     poly2_dict = dict(poly_dict)
     coefficient_dict = poly_dict['parameters'][0]
     poly2_dict['parameters'] = []
-    poly2_dict['_id_name'] = "eab91275-88db-4855-917a-cdcbe7209592"
     for i, coeff in enumerate(coeff_list):
         param_dict = dict(coefficient_dict)
         param_dict['_id_name'] = 'a{}'.format(coeff)