+++ hyperspy/_components/voigt.py
@@ -28,7 +28,6 @@ sigma2fwhm = 2 * math.sqrt(2 * math.log(2))
 
 
 class Voigt(Expression):
-
     r"""Voigt component.
 
     Symmetric peak shape based on the convolution of a Lorentzian and Normal
@@ -84,11 +83,15 @@ class Voigt(Expression):
     doi:10.1107/S0021889886089999
     """
 
-    def __init__(self, centre=10., area=1., gamma=0.2, sigma=0.1,
-                 module=["numpy", "scipy"], **kwargs):
-        # Not to break scripts once we remove the legacy Voigt
-        if "legacy" in kwargs:
-            del kwargs["legacy"]
+    def __init__(
+        self,
+        centre=10.0,
+        area=1.0,
+        gamma=0.2,
+        sigma=0.1,
+        module=["numpy", "scipy"],
+        **kwargs
+    ):
         # We use `gamma_` internally to workaround the use of the `gamma`
         # function in sympy
         super().__init__(
@@ -108,9 +111,9 @@ class Voigt(Expression):
         )
 
         # Boundaries
-        self.area.bmin = 0.
-        self.gamma.bmin = 0.
-        self.sigma.bmin = 0.
+        self.area.bmin = 0.0
+        self.gamma.bmin = 0.0
+        self.sigma.bmin = 0.0
 
         self.isbackground = False
         self.convolved = True
@@ -144,7 +147,7 @@ class Voigt(Expression):
         Examples
         --------
 
-        >>> g = hs.model.components1D.Voigt(legacy=False)
+        >>> g = hs.model.components1D.Voigt()
         >>> x = np.arange(-10, 10, 0.01)
         >>> data = np.zeros((32, 32, 2000))
         >>> data[:] = g.function(x).reshape((1, 1, 2000))
@@ -156,8 +159,9 @@ class Voigt(Expression):
         """
         super()._estimate_parameters(signal)
         axis = signal.axes_manager.signal_axes[0]
-        centre, height, sigma = _estimate_gaussian_parameters(signal, x1, x2,
-                                                              only_current)
+        centre, height, sigma = _estimate_gaussian_parameters(
+            signal, x1, x2, only_current
+        )
         scaling_factor = _get_scaling_factor(signal, axis, centre)
 
         if only_current is True:
@@ -170,14 +174,14 @@ class Voigt(Expression):
         else:
             if self.area.map is None:
                 self._create_arrays()
-            self.area.map['values'][:] = height * sigma * sqrt2pi
+            self.area.map["values"][:] = height * sigma * sqrt2pi
             if axis.is_binned:
-                self.area.map['values'][:] /= scaling_factor
-            self.area.map['is_set'][:] = True
-            self.sigma.map['values'][:] = sigma
-            self.sigma.map['is_set'][:] = True
-            self.centre.map['values'][:] = centre
-            self.centre.map['is_set'][:] = True
+                self.area.map["values"][:] /= scaling_factor
+            self.area.map["is_set"][:] = True
+            self.sigma.map["values"][:] = sigma
+            self.sigma.map["is_set"][:] = True
+            self.centre.map["values"][:] = centre
+            self.centre.map["is_set"][:] = True
             self.fetch_stored_values()
             return True