+++ hyperspy/component.py
@@ -22,10 +22,8 @@ import traits.api as t
 from traits.trait_numeric import Array
 import sympy
 from sympy.utilities.lambdify import lambdify
-from packaging.version import Version
 from pathlib import Path
 
-import hyperspy
 from hyperspy.misc.utils import slugify
 from rsciio.utils.tools import (incremental_filename,
                                     append2pathname,)
@@ -1254,9 +1252,11 @@ class Component(t.HasTraits):
         """
 
         if dic['_id_name'] == self._id_name:
+            # Restoring of polynomials saved with Hyperspy <v1.5
             if (self._id_name == "Polynomial" and
-                    Version(hyperspy.__version__) >= Version("2.0")):
-                # in HyperSpy 2.0 the polynomial definition changed
+                    dic['parameters'][0]['_id_name'] == "coefficients"):
+                # in HyperSpy v1.5 the polynomial definition changed
+                # and the legacy option was removed in v2.0
                 from hyperspy._components.polynomial import convert_to_polynomial
                 dic = convert_to_polynomial(dic)
             load_from_dictionary(self, dic)