+++ hyperspy/models/edsmodel.py
@@ -298,8 +298,7 @@ class EDSModel(Model1D):
         order: int
             The order of the polynomial
         """
-        with ignore_warning(message="The API of the `Polynomial` component"):
-            background = create_component.Polynomial(order=order, legacy=False)
+        background = create_component.Polynomial(order=order)
         background.name = 'background_order_' + str(order)
         background.isbackground = True
         self.append(background)